Properly namespace Combat scripts
