+++ Assets/Scripts/Combat/Events/UnitLives.cs
@@ -1,4 +1,5 @@
 ﻿using GameEvents;
+using Combat.Units;
 
 namespace Combat.Events {
     public class UnitLives : GameEvent<Unit> { }
