+++ Assets/Scripts/Combat/UnitGroups/Controller.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 // NOTE(jordan): List
 using System.Collections.Generic;
+using Combat.Units;
 
 namespace UnitGroups {
     /// <summary>
